Added a DelayFilter to prevent sending messages faster than allowed.
The Toggl API has some rate limits built in. I was receiving HTTP 429
(too many requests) when I ran the full suite, which indicated the tests
were running fast enough that we hit the rate limit on requests before
they finished. Also refactored some of the constants to make them easier
to adjust in case the paths change in the API, and added a check for
premium flag on the workspace for working with a Task (premium feature
only.

Added destroyProject, and added call to it in the JUnit afterClass().